Consolidate CAA functions into va/validation-authority.go and core/dns.go
