+++ va/validation-authority.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/letsencrypt/boulder/core"
 	blog "github.com/letsencrypt/boulder/log"
+	"github.com/letsencrypt/boulder/policy"
 )
 
 // ValidationAuthorityImpl represents a VA
@@ -368,11 +369,90 @@ func (va ValidationAuthorityImpl) UpdateValidations(authz core.Authorization, ch
 	return nil
 }
 
+// CAA Holds decoded CAA record.
+type CAA struct {
+	flag     uint8
+	tag      string
+	value    string
+	valueBuf []byte
+}
+
+// CAASet consists of filtered CAA records
+type CAASet struct {
+	Issue     []*dns.CAA
+	Issuewild []*dns.CAA
+	Iodef     []*dns.CAA
+	Unknown   []*dns.CAA
+}
+
+// returns true if any CAA records have unknown tag properties and are flagged critical.
+func (caaSet CAASet) criticalUnknown() bool {
+	if len(caaSet.Unknown) > 0 {
+		for _, caaRecord := range caaSet.Unknown {
+			// Critical flag is 1, but according to RFC 6844 any flag other than
+			// 0 should currently be interpreted as critical.
+			if caaRecord.Flag > 0 {
+				return true
+			}
+		}
+	}
+
+	return false
+}
+
+// Filter CAA records by property
+func newCAASet(CAAs []*dns.CAA) *CAASet {
+	var filtered CAASet
+
+	for _, caaRecord := range CAAs {
+		switch caaRecord.Tag {
+		case "issue":
+			filtered.Issue = append(filtered.Issue, caaRecord)
+		case "issuewild":
+			filtered.Issuewild = append(filtered.Issuewild, caaRecord)
+		case "iodef":
+			filtered.Iodef = append(filtered.Iodef, caaRecord)
+		default:
+			filtered.Unknown = append(filtered.Unknown, caaRecord)
+		}
+	}
+
+	return &filtered
+}
+
+func (va *ValidationAuthorityImpl) getCAASet(domain string, dnsResolver *core.DNSResolver) (*CAASet, error) {
+	domain = strings.TrimRight(domain, ".")
+	splitDomain := strings.Split(domain, ".")
+	// RFC 6844 CAA set query sequence, 'x.y.z.com' => ['x.y.z.com', 'y.z.com', 'z.com']
+	for i := range splitDomain {
+		queryDomain := strings.Join(splitDomain[i:], ".")
+		// Don't query a public suffix
+		if _, present := policy.PublicSuffixList[queryDomain]; present {
+			break
+		}
+
+		// Query CAA records for domain and its alias if it has a CNAME
+		for _, alias := range []bool{false, true} {
+			CAAs, err := va.DNSResolver.LookupCAA(queryDomain, alias)
+			if err != nil {
+				return nil, err
+			}
+
+			if len(CAAs) > 0 {
+				return newCAASet(CAAs), nil
+			}
+		}
+	}
+
+	// no CAA records found
+	return nil, nil
+}
+
 // CheckCAARecords verifies that, if the indicated subscriber domain has any CAA
 // records, they authorize the configured CA domain to issue a certificate
 func (va *ValidationAuthorityImpl) CheckCAARecords(identifier core.AcmeIdentifier) (present, valid bool, err error) {
 	domain := strings.ToLower(identifier.Value)
-	caaSet, err := getCaaSet(domain, va.DNSResolver)
+	caaSet, err := va.getCAASet(domain, va.DNSResolver)
 	if err != nil {
 		return
 	}
